Registration integrated
