+++ src/utils/isSSR.ts
@@ -1,0 +1,6 @@
+/**
+ * Validate Server Side Rendering
+ *
+ * @returns validation boolean
+ */
+export default () => typeof window === 'undefined';
